Cloud-Operations-GCP/collection_cloud_function/requirements.txt
requests
beautifulsoup4
google-cloud-storage
flask
